Refractoring Finished
The refractoring is done. I surely made the code more concise and
robust.
character is a class, an array of characters is what holds the info for
the HTML elements, npc and isHighlighted neatly organized. Some
functions migrated to the class and became methods, some were completely
rewritten (charDel, clearRound, etc.)