Random Added
A bunch of functions are added. Most of them handle automated
randomizing of the initiative number for NPCs. There is also an
animation function that unrolls and rolls up the randomize button. Also,
a couple of fields for character are added. They involve holding the
information needed for the randomizing.