Security Check In Progress
I wrote the function isInputUnique(). It is yet to be finished but so
far, I have found that accessing the names of the characters directly
from the array of objects is twice as efficient as putting them all in a
single string and looking for a match to check how unique it is. The
update is not finished.