Security Check In Progress v0.3
The function isInputUnique() is now integrated into the check before the
character is added as an object. Now, I need to make animateBorder()
function more versatile. It needs to take an array of elements that are
to be highlighted along with the respective CSS properties that need to
change colors (blink.)