Security Check in Progress v0.4
I made the animate() function return the original colour to the object's
property after it is done blinking it. Then, I made the character's name
(charName - CSS) blink. Now, if the text in the name field is identical
with already existing name, the name field's border and the text of the
already existing name will blink. However, I noticed that
charAddNpcInitMaster (CSS version of the naming) is not unrolling
horizontally upon selection of the NPC option in the CharNPCSelect
(again, CSS.) The bug is to be fixed.