Security Check in Progress v6.0
I fixed bugs where player characters started to inherit the colour of
NPC after they blinked indicating that the same-name character was about
to be added. Also, I straightened up the code in the checkForm(). It
still needs some refractoring (potentially, into a single if statement)
but it works well for now.