Fix 0.1
In this minor update, I started doing the options feature set. It went
as far as putting "Options:" into the HTML code.
On the other side, I added the "Delete All" button. It calls deleteAll()
which in turn calls deleteChar() for every caracter in chars[] array. In
order to make deleteChar() work properly with such frequent calls, I had
to alter it and now it works as intended.
Also, I moved the calling of animate() function for the initiative
fields of NPC into a separate function called rollNPCInitContainer(). It
used to have math functions in the parameters where the function was
called (in the HTML options' onchange.) With the update, it fixed the
bug that a user was encountering on his Chrome. I thought I was the only
one with that problem so I did not care for it much. From now on, I
should fix bugs on individual level too.