Progress on Fix 0.2
There is a certain progress done in this update.
First of all, a charInitInput is rapidly checked for validity (using
checkInitInput() and checkInputText()). If the validation is not passed,
the failed fields will animateBorder() and the findMax()'s further
calculations will not happen.
Delete All button is added. It deletes all the characters in the scene
using deleteAll() function.
checkInputText() is extended to specify if an empty field is should be
accepted or not.
FindMax() now does a charInitInput check, as previously mentioned.
Cookie functions are to be completely rewritten. The idea is to store
three cookies with predefined names: tools, the current scene, the saved
scenes. Some of the functions are implemented in this progress save:
toolsSave() stores tools and options into the cookies
foolsLoad() gets the settings and tools from the cookies. Perhaps, I
should look into error handling on this one.
toolsToString() same as sceneToString, same as any toString functions.
cookieSet() a unified function which I use to store the cookies. It
takes the cookie name and the cookie value as parameters and stores them
for one year.
cookieGet() gets a cookie with the specified name. If a cookie is not
found, returns null.
With the recent development of cookies, I will avoid the problem of the
program fetching all the cookies stored for the whole website. Looking
for a specifically-named cookies (eg: "SIT_tools", "SIT_currentScene",
"SIT_scenes") should help.