Progress on Fix 0.2 v1.1
I completed the pulseBorder() function. It takes an element, it's
property in the style, and the property's value - the original color -
and makes it go up with the red channel and back to its original color.
The interesting thing is that the frame duration is calculated so that
regardless of how low the red channel originally is, it takes 0.5 second
to reach its "FF" state and exactly as much to get back to its original
value.