Dev. Update 1.4.0 v0.3
I have fixed the previously occurring bug with no-name single character
showing up. The reason why that happened is because it tried to load the
SIT_currentScene from cookies but when the previous scene was under
deletion, the SIT_currentScene was changing accordingly, after every
character deleted, eventually resulting in a "" string. The string was
then saved into cookies under SIT_currentScene, and then immediately
taken out to be put into scenePopulate() (see the algorithm of
loadScene().) This is why the first and the only character that appeared
in the scene browser (that's a cool name, I'll keep calling the list of
characters that) was the "void" character.
It is fixed now as sceneLoad() checks if the length of the string passed
into scenePopulate is more than 0 - and if it is not - does not call the
function.
currentSceneLoad() is complete. Now, every time the page loads,
currentSceneLoad() takes "SIT_currentScene" out of cookies and puts it
into scenePopulate().
Also, a delayed info save bug is fixed. When we changed the content of
some initInput, the field called initInputChanged() onKeyPress which in
turn saved the scene with currentSceneSave() and then called
checkInitInput(). In checkInitInput() is a delay before the actual check
of the new content put in place to wait until the field changes its
value according to the key pressed. Since the scene was saved in
initInputChanged(), before the value chenged, currentSceneSave() was
saving the previous content, before the button pushed had its effect on
the value of the field.
To fix this problem, I moved the delay into initInputChenged() function.
It seems more appropriate there not only because now currentSceneSave()
waits until the change is applied but also because when checkInitInput()
is called from findMax() (to check the validity before the next phase,)
it does not benefit from waiting because (most likely) no button
triggering the initInput.value change was pushed within the last
millisecond.
The next big step: findMaxS() - the standard rule system.