Dev. Update 1.4.0 v0.5
The tools save bug is fixed now. It was because the lines that assigned
the radiobuttons to their state (lines 259 and 260) were reading only
the first character of the "true" or "false" string. That happened
because save.split() apparently does not produce an array if there is
only one element to put into it.
Next in line: standart rule engine or the design.