Dev. Update 1.4.0 v0.6
Fixes:
-Now, you can switch to the next charInitInput with the Tab button.
-The character is added upon Enter (I returned the <form> to its place)
+Also, I need to add <form> to the rest of input clusters (scene and
scene to text)
-Now, when you tab through the charAddForm's elements and "Player" is
selected in the charNpcSelect, you do not tab through
charNpcInitBaseInput and charNpcInitBaseD6. They turn visible only when
you select NPC in the options.
-Once nextButton is pressed, it is not active for 2 seconds. Changes are
made in the nextPhase() function. The button sometimes is double-clicked
and there is no "undo" function (yet?)
+I could also add an option so the users will be able to select whether
they want the inactive period or not.

Next:
-add standard rule system
-rebuild the interface (HTML and CSS,) add position: relative to the
charContainer
-sort the charContainers by the charInitInput.value