new UI 0.5.3
+The bug is fixed. As I suspected, adding innerHTML and appendChild()
make whatever you appending not recognizable (at least, that's what
happened with the pc button.)
+I included all the HTML text added by charAdd() into their own
tespective divs. It had solved the problem.
+Added formBrowserContainer to store the forms solely. Its indefinite
height will allow to put the "Add" buttons right after the list.
+Now, in order for a new form to appear, at least all but one form must
be filled (aka, (numOfForms-1).)
-Implement "Add to Scene" button. It will add characters to the current
scene, just like before.
-Implement "Add to Group" button. It will add the characters to a new or
the existing group of goons that can later be added to the scene with a
click of a button (or something: either a button per group on the top
panel or a select field with the groups in it.
--The UI of existing groups should include the name of the group and the
first 3 members under it (with a "..." followed if more than 3.)
-Implement team-enemies distinction.